backend/torchimg/requirements.txt
albucore               
albumentations        
annotated-types        
anyio                  
blinker                
click                 
colorama              
contourpy              
cycler                 
dep-logic             
distlib                
exceptiongroup         
fastapi                
filelock               
findpython            
fonttools              
fsspec                
greenlet               
h11                    
h5py                   
hishel                 
httpcore               
httpx                  
id                     
idna                   
installer              
Jinja2                 
kiwisolver             
MarkupSafe             
matplotlib             
mpmath                 
msgpack                
networkx              
numpy                  
opencv-contrib-python  
opencv-python          
opencv-python-headless 
packaging              
pbs-installer          
pdm                    
pillow                
pip                    
platformdirs           
pydantic               
pydantic_core          
pyparsing              
pyproject_hooks        
python-dateutil        
python-multipart      
PyYAML                
resolvelib             
scipy                  
setuptools           
simsimd              
six                 
sniffio            
socksio               
SQLAlchemy             
sqlmodel               
starlette             
stringzilla            
sympy                  
terminaltables         
tomli                  
tomlkit                
torch                
torchvision            
tqdm                  
truststore             
typing_extensions      
typing-inspection      
unearth                
uvicorn               
virtualenv             
watchfiles             
wheel                  
mysql

